feat!: add tailwind, new auth logic
Now if you're not signed in, it automatically sends you to the sign in
page. This is good because it isn't sending a state error when you
update global redux state and redirect at the same time.

Tailwind is now installed, haven't tested to see if it works yet, but
the app renders. Using the recommended installation from Tailwind's
Official site.